@@ -1,8 +1,10 @@
 package com.example.xyzreader.data;
 
 import android.content.Context;
-import android.content.CursorLoader;
 import android.net.Uri;
+
+import androidx.loader.content.CursorLoader;
+
 
 /**
  * Helper for loading a list of articles or a single article.
